Feature

+++ ArtworkSharing/Controllers/CommentController.cs
@@ -1,6 +1,7 @@
 ﻿using ArtworkSharing.Core.Interfaces.Services;
 using ArtworkSharing.Core.ViewModels.Comments;
 using Microsoft.AspNetCore.Mvc;
+using System.Security.Claims;
 
 namespace ArtworkSharing.Controllers;
 
@@ -24,7 +25,7 @@ public class CommentController : ControllerBase
     public async Task<IActionResult> GetCommentByArtworkId([FromRoute] Guid id)
     {
         if (id == Guid.Empty) return BadRequest(new { Message = "Not found artwork" });
-        var c = await _commentService.GetCommentByArtworkId(id);
+
         return Ok(await _commentService.GetCommentByArtworkId(id));
     }
 
@@ -36,16 +37,14 @@ public class CommentController : ControllerBase
     [HttpPost]
     public async Task<IActionResult> CreateComment(CreateCommentModel createCommentModel)
     {
-        if (HttpContext.Items.TryGetValue("UserId", out var u) is false)
-        {
-            return BadRequest();
-        }
+        var uidClaim = User.Claims.FirstOrDefault(c => c.Type == ClaimTypes.NameIdentifier);
+        Guid uid = new Guid(uidClaim!.Value);
 
-        if (u == null) return BadRequest();
+        if (uid == Guid.Empty) return Unauthorized();
 
         if (createCommentModel == null) return BadRequest();
 
-        var rs = await _commentService.Add(createCommentModel.ArtworkId, Guid.Parse(u + ""), createCommentModel.Content);
+        var rs = await _commentService.Add(createCommentModel.ArtworkId, uid, createCommentModel.Content);
         return rs != null!
             ? StatusCode(StatusCodes.Status201Created, rs)
             : StatusCode(StatusCodes.Status500InternalServerError, new { Message = "Create failed" });
@@ -59,6 +58,17 @@ public class CommentController : ControllerBase
     [HttpDelete("{id}")]
     public async Task<IActionResult> DeleteComment([FromRoute] Guid id)
     {
+        var uidClaim = User.Claims.FirstOrDefault(c => c.Type == ClaimTypes.NameIdentifier);
+        Guid uid = new Guid(uidClaim!.Value);
+
+        if (uid == Guid.Empty) return Unauthorized();
+
+        var comment = await _commentService.GetOne(id);
+
+        if (comment == null) return BadRequest();
+
+        if (comment.CommentedUserId != uid) return BadRequest();
+
         if (id == Guid.Empty) return BadRequest(new { Message = "Not found comment" });
         return await _commentService.Delete(id)
             ? StatusCode(StatusCodes.Status204NoContent)
@@ -73,6 +83,17 @@ public class CommentController : ControllerBase
     [HttpPut]
     public async Task<IActionResult> UpdateComment(UpdateCommentModel updateCommentModel)
     {
+        var uidClaim = User.Claims.FirstOrDefault(c => c.Type == ClaimTypes.NameIdentifier);
+        Guid uid = new Guid(uidClaim!.Value);
+
+        if (uid == Guid.Empty) return Unauthorized();
+
+        var comment = await _commentService.GetOne(updateCommentModel.Id);
+
+        if (comment == null) return BadRequest();
+
+        if (comment.CommentedUserId != uid) return BadRequest();
+
         if (updateCommentModel == null || updateCommentModel.Id == Guid.Empty) return BadRequest();
         var rs = await _commentService.Update(updateCommentModel);
         return rs != null!
